+++ src/holiday/holiday.module.ts
@@ -3,7 +3,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { ScheduleModule } from '@nestjs/schedule';
 import { RotationHolidayEntity } from './entity/holiday.entity';
 import { HolidayService } from './holiday.service';
-import { HolidayRepository } from './holiday.repository';
+import { HolidayRepository } from './repository/holiday.repository';
 
 @Module({
   imports: [
