+++ src/rotation/rotations.module.ts
@@ -5,10 +5,10 @@ import { RotationEntity } from './entity/rotation.entity';
 import { RotationAttendeeEntity } from './entity/rotation-attendee.entity';
 import { RotationsService } from './rotations.service';
 import { RotationsController } from './rotations.controller';
-import { CustomRotationRepository } from './rotations.repository';
+import { CustomRotationRepository } from './repository/rotations.repository';
 import { RotationHolidayEntity } from '../holiday/entity/holiday.entity';
 import { HolidayService } from '../holiday/holiday.service';
-import { HolidayRepository } from '../holiday/holiday.repository';
+import { HolidayRepository } from '../holiday/repository/holiday.repository';
 import { HolidayModule } from '../holiday/holiday.module';
 import { UserService } from 'src/user/user.service';
 import { UserRepository } from 'src/user/repository/user.repository';
